move caching from extensions folder to basic source dir
