+++ fuglu/tests/unit/caching_test.py
@@ -1,5 +1,5 @@
 import unittest
-from fuglu.extensions.caching import smart_cached_memberfunc, smart_cached_property, CacheLimits, CacheStats
+from fuglu.caching import smart_cached_memberfunc, smart_cached_property, CacheLimits, CacheStats
 
 class CachingTests(unittest.TestCase):
     def test_cachedProperty(self):
